kubeadm: fix a bug where file copy(backup) could not be executed correctly on Windows platform during upgrade
