+++ cmd/kubeadm/app/cmd/phases/util.go
@@ -17,8 +17,6 @@ limitations under the License.
 package phases
 
 import (
-	"os"
-
 	"k8s.io/component-base/version"
 
 	kubeadmapiv1 "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta3"
@@ -32,18 +30,4 @@ func SetKubernetesVersion(cfg *kubeadmapiv1.ClusterConfiguration) {
 		return
 	}
 	cfg.KubernetesVersion = version.Get().String()
-}
-
-// CopyFile copy file from src to dest.
-func CopyFile(src, dest string) error {
-	fileInfo, err := os.Stat(src)
-	if err != nil {
-		return err
-	}
-	contents, err := os.ReadFile(src)
-	if err != nil {
-		return err
-	}
-	err = os.WriteFile(dest, contents, fileInfo.Mode())
-	return err
 }
